+++ CHANGELOG.md
@@ -25,3 +25,4 @@
 ##### v0.4 [14.03.2021]: 
 * (ERICK) Added javadoc documentation to RoomHandler.
 * (ERICK) Created `ModuleHandler` class.
+* (ERICK) Created `refreshModuleHandler()` method.